rename regex.js to rules.js
